backend/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
supabase>=2.0.0
python-dotenv>=1.0.0
gpxpy>=1.6.0
click>=8.0.0
numpy>=1.24.0
haversine>=2.8.0
pillow>=10.0.0
pillow-heif>=0.13.0
piexif>=1.1.3
supervision>=0.16.0